Week_7/github-agent/requirements.txt
mcp>=1.0.0
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
websockets>=12.0
pydantic>=2.0
python-dotenv>=1.0
rich>=13.0
